+++ src/components/search/data/tests/constants.js
@@ -1,0 +1,14 @@
+export const SUBJECTS = {
+  COMPUTER_SCIENCE: 'Computer Science',
+  COMMUNICATION: 'Communication',
+};
+
+export const AVAILABLILITY = {
+  AVAILABLE_NOW: 'Available Now',
+  UPCOMING: 'Upcoming',
+};
+
+export const FACET_ATTRIBUTES = {
+  AVAILABLILITY: 'availability',
+  SUBJECTS: 'subjects',
+};
