+++ src/components/search/data/tests/utils.test.js
@@ -1,4 +1,4 @@
-import { SUBJECTS } from '../constants';
+import { SUBJECTS } from './constants';
 import {
   sortItemsByLabelAsc,
   updateRefinementsFromQueryParams,
